# improve vec_hash performance