--- a/polars/polars-arrow/src/lib.rs
+++ b/polars/polars-arrow/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod array;
 pub mod bit_util;
+pub mod bitmap;
 pub mod builder;
 pub mod error;
 pub mod is_valid;