```
Gnuplot not found, using plotters backend
bench vec hash null rate: 0%                                                                            
                        time:   [64.121 us 64.133 us 64.149 us]
                        change: [-64.831% -64.661% -64.479%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) high mild
  10 (10.00%) high severe

bench vec hash null rate: 20%                                                                            
                        time:   [73.032 us 73.165 us 73.330 us]
                        change: [-72.295% -72.236% -72.171%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 23 outliers among 100 measurements (23.00%)
  6 (6.00%) low severe
  4 (4.00%) low mild
  6 (6.00%) high mild
  7 (7.00%) high severe

bench vec hash null rate: 50%                                                                            
                        time:   [81.263 us 81.345 us 81.414 us]
                        change: [-78.007% -77.995% -77.984%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  8 (8.00%) low severe
  2 (2.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe

bench vec hash null rate: 80%                                                                            
                        time:   [74.022 us 74.117 us 74.218 us]
                        change: [-70.255% -70.109% -69.970%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

Benchmarking bench vec hash combine null rate: 0%: Collecting 100 samples in estimated 5.1689 s (61k iterati                                                                                                            bench vec hash combine null rate: 0%                        
                        time:   [85.254 us 85.282 us 85.316 us]
                        change: [-63.589% -63.402% -63.284%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low severe
  4 (4.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe

Benchmarking bench vec hash combine null rate: 20%: Collecting 100 samples in estimated 5.1072 s (20k iterat                                                                                                            bench vec hash combine null rate: 20%                        
                        time:   [252.55 us 252.95 us 253.35 us]
                        change: [-18.249% -17.611% -16.926%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

Benchmarking bench vec hash combine null rate: 50%: Collecting 100 samples in estimated 5.0704 s (15k iterat                                                                                                            bench vec hash combine null rate: 50%                        
                        time:   [333.56 us 333.76 us 333.99 us]
                        change: [-18.121% -18.058% -17.982%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

Benchmarking bench vec combine hash null rate: 80%: Collecting 100 samples in estimated 5.3235 s (25k iterat                                                                                                            bench vec combine hash null rate: 80%                        
                        time:   [210.59 us 210.63 us 210.66 us]
                        change: [-24.217% -24.190% -24.160%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  2 (2.00%) low mild
  8 (8.00%) high mild
  4 (4.00%) high severe


```